refactor: extract shared HTTP handlers into internal/httpserver
Eliminate ~460 lines of duplication between server.go and mobile/mobile.go.
All route handlers now live in internal/httpserver with an Options struct
to configure auth, history, and resource serving. Both entry points are
reduced to thin wrappers.
